[vioscsi] Define MAX_PHYS_SEGMENTS to the "true limit"
It is the only sensible thing to do when we allow user to set the
number of physical breaks to a lower value while also want to keep
the +1 practice as default / much as possible, and when we need to
make sure the ultimate value is clamped to it.

We don't want to do +1 for any user input values as that only leads
to (more) confusion/surprise, given that qemu is already doing the
probably-no-longer-necessary -2 adjustment. It can hardly be sane
to try to optimize a value when it can be entered arbitrarily (so
same story with the registry approach).

This commit has the assumption that the original VIRTIO_MAX_SG was
derived from HEAD + MAX_PHY_SEGMENTS(before/no +1) + 1 + TAIL.

Signed-off-by: Tom Yan <tom.ty89@gmail.com>