vioscsi: accept (virt)queue_size as source of max_physical_breaks
qemu provides the (virt)queue_size option for user to set seg_max.
accept it as an alternative to the registry value.

As qemu reduce the input/default value by 2 before setting seg_max
to that, we compensate it here as we already reserved for the head
and tail (and the "DMA plus one") when we define VIRTIO_MAX_SG.

While the default value will either be 126 / 254, it is clamped to
MAX_PHYS_SEGMENTS (after + 2) when we take it, so practically it
can only be used to lower max_physical_breaks. In other words, the
default max_physical_breaks remains unchanged.